==> data_prep/07_process_fish_stocks.Rmd <==
---
title: "Create stock rasters"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
---


```{r setup, message=F, warning=F}
knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',message = FALSE, warning = FALSE)

library(raster)
library(tidyverse)
library(sf)
library(rredlist)
library(furrr)
library(doParallel)
library(parallel)
library(foreach)
library(paletteer)

source(here::here("common.R"))
```

```{r}
stocks <- read_rds(file.path(emLab_project_dir, "data", "02_processed", "food_provision", "Cleanmegacell.rds"))

cells <- read_rds(file.path(emLab_project_dir, "data", "02_processed", "food_provision", "CleanCoordmegacell.rds"))

stocks_distributions <- bind_cols(stocks, cells) %>% 
  select(lon, lat, everything()) %>% 
  gather(stock_id, suitability, -lon, -lat)
```

# Generate stock rasters

```{r}
empty_raster <- raster(resolution = 0.5)

registerDoParallel(50) #register cores

unique_stock_ids <- unique(stocks_distributions$stock_id)

foreach(stock = unique_stock_ids) %dopar% { 
  
  stock_data <- stocks_distributions %>% 
    filter(stock_id == "Fis-27719",
           suitability > 0) %>% 
    select(lon, lat, suitability)
  
  cells <- cellFromXY(empty_raster, as.matrix(stock_data[,1:2]))
  
  stock_raster <- empty_raster
  
  stock_raster[cells] <- pull(stock_data, suitability)
  
  stock_raster <- projectRaster(stock_raster, ocean_low_res_moll)
  
  stock_raster <- mask(stock_raster, ocean_low_res_moll)
  
  raster::writeRaster(stock_raster, 
                      file.path(emLab_project_dir, "data", "02_processed", "stock_distributions", paste0("Fis-27719", ".tif")),
                      overwrite = T)
}
```
