==> data_prep/10_process_admin_layers.Rmd <==
---
title: "Current highly protected MPAs"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
---

```{r setup, message=F,warning=F}
knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',message = FALSE, warning = FALSE)

library(sf)
library(raster)
library(tidyverse)
library(rredlist)
library(furrr)
library(rgdal)
library(parallel)
library(foreach)
library(doParallel)

source(here::here("common.R"))
```

# Create MPAs raster

```{r}
# The input file geodatabase

MPAtlas_gdb_path <- file.path(emLab_shrd_data_dir, "mpa-atlas", "MPAtlas_20190910.gdb/")

# List all feature classes in a file geodatabase

fc_list <- ogrListLayers(MPAtlas_gdb_path)

#print(fc_list)

# Read the feature class

all_MPAs <- sf::st_read(dsn = MPAtlas_gdb_path, layer = "mpatlas_20190910_poly")

mpas_info <- all_MPAs %>% 
  st_drop_geometry()
```

## implememted and highly protected

```{r}
fully_highly_protected_MPAs <- all_MPAs %>% 
  filter(is_mpa == TRUE,  
         status != "Proposed", 
         implemented == TRUE,
         protection_level %in% c("Full, High", "Full, High (Partially Zoned)"))

#colnames(fully_highly_protected_MPAs)

#sf::st_write(fully_highly_protected_MPAs, "../../../gcs/spatial-datasets/MPAtlas_20190910/fully_and_highly_protected_implemented_MPAs.shp", delete_layer=TRUE)

# fully_highly_protected_MPAs_raster_moll <- fully_highly_protected_MPAs %>% 
#   st_transform("+proj=moll") %>% 
#   fasterize::fasterize(ocean_low_res_moll) %>% 
#   mask(ocean_low_res_moll)

# fully_highly_protected_MPAs_map <- tmap::tm_shape(fully_highly_protected_MPAs_raster_moll)+
#     tmap::tm_raster(title = "",
#                     palette = c("green"),
#                     style = "cont")+
#     tmap::tm_shape(land_shp_moll)+
#     tmap::tm_polygons(col = "black", border.col = "transparent")+
#     tmap::tm_layout(title = "Implemented Fully/Highly Protected MPAs",
#                     title.position = c("center", .95),
#                     inner.margins = c(0.12, 0.04, 0.08, 0.04), 
#                     frame = F)
# 
# tmap::tmap_save(fully_highly_protected_MPAs_map,
#                 filename = file.path(DIR_GCS, "masks/figures/fully_highly_protected_implemented_MPAs.png"),
#                 dpi = 300,  height = 4, width = 8 )
# 
# writeRaster(fully_highly_protected_MPAs_raster_moll, 
#             file.path(DIR_GCS,"masks/data/processed/fully_highly_protected_implemented_MPAs.tif"),
#             overwrite = T)
```

## designated & highly protected

```{r}
designated_fully_highly_protected_MPAs <- all_MPAs %>% 
  filter(is_mpa == TRUE,  
         status == 'Designated',  
         protection_level %in% c("Full, High", "Full, High (Partially Zoned)"))

# designated_fully_highly_protected_MPAs_raster_moll <- designated_fully_highly_protected_MPAs %>% 
#   st_transform("+proj=moll") %>% 
#   fasterize::fasterize(ocean_low_res_moll) %>% 
#   mask(ocean_low_res_moll)
# 
# designated_fully_highly_protected_MPAs_map <- tmap::tm_shape(designated_fully_highly_protected_MPAs_raster_moll)+
#     tmap::tm_raster(title = "",
#                     palette = c("green"),
#                     style = "cont")+
#     tmap::tm_shape(land_shp_moll)+
#     tmap::tm_polygons(col = "black", border.col = "transparent")+
#     tmap::tm_layout(title = "Designated Fully/Highly Protected MPAs",
#                     title.position = c("center", .95),
#                     inner.margins = c(0.12, 0.04, 0.08, 0.04), 
#                     frame = F)
# 
# tmap::tmap_save(designated_fully_highly_protected_MPAs_map,
#                 filename = file.path(DIR_GCS, "masks/figures/fully_highly_protected_designated_MPAs.png"),
#                 dpi = 300,  height = 4, width = 8 )
# 
# writeRaster(designated_fully_highly_protected_MPAs_raster_moll, 
#             file.path(DIR_GCS,"masks/data/processed/fully_highly_protected_designated_MPAs.tif"), 
#             overwrite = T)
```

## manual review

Which are designated but not implemented according to MPAtlas?

```{r}
implemented_MPA_ids <- fully_highly_protected_MPAs %>% 
  st_set_geometry(NULL) %>% 
  distinct(mpa_id) %>% 
  pull()

(designated_fully_highly_protected_MPAs %>% 
   st_set_geometry(NULL) %>% 
   filter(!mpa_id %in% implemented_MPA_ids))
     #write_csv("mpas_to_review.csv"))
```

```{r}
MPAs_to_xclude <- c(68808636, 68808621, 68807893, 15625) # Rapa Nui Rahui	, Islas del Pacífico de la Península de Baja California Core Zone, Ascension, Natural Park of the Coral Sea	, 

MPAs_to_add <- designated_fully_highly_protected_MPAs %>% 
  st_set_geometry(NULL) %>% 
  filter(!mpa_id %in% implemented_MPA_ids,
         !mpa_id %in% MPAs_to_xclude)

# c(68808472, 68808467, 68808466, 68808416, 68808415, 68808409, 15907, 14985, 14982, 12864, 9234, 737)
```

```{r}
reviewed_fully_highly_protected_MPAs_moll <- all_MPAs %>% 
  filter((is_mpa == TRUE & status != "Proposed" & implemented == TRUE & protection_level %in% c("Full, High", "Full, High (Partially Zoned)")) | 
           mpa_id %in% MPAs_to_add$mpa_id | country == "GAB" ) %>% 
  filter(mpa_id != 68808627) %>% # remove Pacifico mexicano profundo
  st_transform("+proj=moll") 

reviewed_fully_highly_protected_MPAs_raster_moll <- reviewed_fully_highly_protected_MPAs_moll %>% 
  fasterize::fasterize(ocean_low_res_moll) %>% 
  mask(ocean_low_res_moll)

reviewed_fully_highly_protected_MPAs_map <- tmap::tm_shape(reviewed_fully_highly_protected_MPAs_raster_moll)+
    tmap::tm_raster(title = "",
                    palette = c("green"),
                    style = "cont")+
    tmap::tm_shape(land_50_moll)+
    tmap::tm_polygons(col = "black", border.col = "transparent")+
    tmap::tm_layout(title = "Reviewed Fully/Highly Protected MPAs",
                    title.position = c("center", .95),
                    inner.margins = c(0.12, 0.04, 0.08, 0.04), 
                    frame = F)

tmap::tmap_save(reviewed_fully_highly_protected_MPAs_map,
                filename = file.path(emLab_project_dir, "figures", "exploratory", "fully_highly_protected_reviewed_MPAs.png"),
                dpi = 300,  height = 4, width = 8 )

writeRaster(reviewed_fully_highly_protected_MPAs_raster_moll, 
            filename = file.path(emLab_project_dir, "data", "02_processed", "masks", "fully_highly_protected_reviewed_MPAs.tif"),
            overwrite = T)

sf::st_write(reviewed_fully_highly_protected_MPAs_moll,
             file.path(emLab_project_dir, "data", "02_processed", "masks", "fully_highly_protected_reviewed_MPAs.gpkg"),
             append = FALSE)
```

```{r}
reviewed_implemented_MPAs_moll <- all_MPAs %>% 
  filter((is_mpa == TRUE & status != "Proposed" & implemented == TRUE) | 
           mpa_id %in% MPAs_to_add$mpa_id | country == "GAB") %>% 
  filter(mpa_id != 68808627) %>% # remove Pacifico mexicano profundo
  st_transform("+proj=moll") 

# tmp_sf <- reviewed_implemented_MPAs_moll %>% 
#   sf::st_simplify(preserveTopology = T, dTolerance = 1000000)
# 
# 
# ggplot()+
#   geom_sf(data = head(tmp_sf, 100), lwd = .2, fill = "transparent")

reviewed_implemented_MPAs_raster_moll <- reviewed_implemented_MPAs_moll %>% 
  fasterize::fasterize(ocean_low_res_moll) %>% 
  mask(ocean_low_res_moll)

reviewed_implemented_MPAs_raster_moll_map <- tmap::tm_shape(reviewed_implemented_MPAs_raster_moll)+
    tmap::tm_raster(title = "",
                    palette = c("green"),
                    style = "cont")+
    tmap::tm_shape(land_50_moll)+
    tmap::tm_polygons(col = "black", border.col = "transparent")+
    tmap::tm_layout(title = "Reviewed implemented MPAs",
                    title.position = c("center", .95),
                    inner.margins = c(0.12, 0.04, 0.08, 0.04), 
                    frame = F)

tmap::tmap_save(reviewed_implemented_MPAs_raster_moll_map,
                filename = file.path(emLab_project_dir, "figures", "exploratory", "implemented_reviewed_MPAs_all_protection_levels.png"),
                dpi = 300,  height = 4, width = 8 )

writeRaster(reviewed_implemented_MPAs_raster_moll, 
            filename = file.path(emLab_project_dir, "data", "02_processed", "masks", "reviewed_implemented_MPAs_all_protection_levels.tif"),
            overwrite = T)
```

# Territories EEZs

```{r}
eez_and_land <- sf::st_read(file.path(emLab_shrd_data_dir, "marine-regions-EEZ-land-union-v3-202003", "EEZ_Land_v3_202030.shp"))

ocean_high_res <- raster(file.path(emLab_shrd_data_dir, "ohi-supplement", "ocean.tif"))

territories_raster_hr <- eez_and_land %>% 
  st_transform(crs(ocean_high_res)) %>% 
  fasterize::fasterize(ocean_high_res, field = "MRGID_TER1", fun = "first") %>% 
  mask(ocean_high_res)

territories_raster_lr <- territories_raster_hr %>% 
  aggregate(fact = 50, fun = "modal")

territories_raster_lr <- territories_raster_lr %>% 
  projectRaster(ocean_low_res_moll, method = "ngb") 

territories_raster_lr <- territories_raster_lr %>% 
  mask(ocean_low_res_moll)

writeRaster(territories_raster_lr, 
            file.path(emLab_project_dir, "data", "02_processed", "masks", "territories.tif"), overwrite = T)

eez_info <- eez_and_land %>% 
  sf::st_drop_geometry()

eez_info %>% 
  janitor::clean_names() %>% 
  rename(name = union) %>% 
  select(mrgid_ter1, iso_ter1, name, mrgid_eez,iso_sov1) %>% 
  mutate_if(is.factor, as.character) %>% 
  write_csv(file.path(emLab_project_dir, "data", "02_processed", "masks", "territories_lookup.csv"))
```
